Support multiple PDDL domains (ROS not tested yet)
